pyproject.toml
[project]
name = "fi-th"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "django>=5.2.3",
    "djangorestframework>=3.16.0",
]
